+++ packages/fluent/src/components/StatusDropdown/StatusColor.tsx
@@ -1,4 +1,4 @@
-import { Dictionary } from "@brhanso/dynamic-forms-core";
+import { Dictionary } from "@bghcore/dynamic-forms-core";
 import React from "react";
 
 interface IStatusColorProps {