+++ packages/fluent/src/fields/HookNumber.tsx
@@ -1,4 +1,4 @@
-import { IHookFieldSharedProps, isNull } from "@brhanso/dynamic-forms-core";
+import { IHookFieldSharedProps, isNull } from "@bghcore/dynamic-forms-core";
 import { ITextFieldProps, TextField } from "@fluentui/react";
 import React from "react";
 import { ReadOnlyText } from "../components/ReadOnlyText";